requirements.txt
attrs
requests
PyYAML